backend/requirements.txt
# LexAI Backend Requirements
# Legal Document Intelligence Platform

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# AI/ML Libraries
torch==2.0.1
transformers==4.33.2
numpy==1.24.3
scikit-learn==1.3.0

# Explainable AI & LLM
shap==0.42.1
lime==0.2.0.1
google-genai==0.3.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# Database
SQLAlchemy==2.0.20

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0

# Development/Testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production Server
gunicorn==21.2.0
